+++ README.md
@@ -21,12 +21,12 @@ See https://github.com/sicpa-dlab/didcomm-demo.
 Example code:
 
 ```python
-from peerdid.keys import Ed25519VerificationKey, X25519KeyAgreementKey
-from peerdid.peer_did import (
+from peerdid.dids import (
     create_peer_did_numalgo_0,
     create_peer_did_numalgo_2,
     resolve_peer_did,
 )
+from peerdid.keys import Ed25519VerificationKey, X25519KeyAgreementKey
 
 encryption_keys = [
     X25519KeyAgreementKey.from_base58(
@@ -83,7 +83,7 @@ Example of DID documents:
     "capabilityDelegation": ["#6MkqRYqQ"]
 }
 
-// did_doc_algo_2
+// did_doc_algo_2_json
 {
     "@context": [
         "https://www.w3.org/ns/did/v1",
