+++ peerdid/__init__.py
@@ -1,9 +1,9 @@
 """Peer DID document generation and resolution."""
 
-from . import core, errors, keys, peer_did
+from . import core, dids, errors, keys
 
 from pydid import DID, DIDDocument
 
 __version__ = "0.4.0"
 
-__all__ = ["__version__", "core", "errors", "keys", "peer_did", "DID", "DIDDocument"]
+__all__ = ["__version__", "core", "errors", "dids", "keys", "DID", "DIDDocument"]
